Refactor css code
