+++ src/overview/onboarding/components/Onboarding.css
@@ -49,52 +49,6 @@
     }
 }
 
-.choiceBtn {
-    padding: 6px 1.6em;
-    border-radius: 2px;
-    font-weight: 600;
-    color: #fff;
-    cursor: pointer;
-    text-decoration: none;
-    background: #777;
-    width: fit-content;
-    border: 2px solid #777;
-    font-size: 24px;
-
-    &:hover,
-    &:active {
-        background-color: #fff;
-        color: #3eb995;
-        border: 2px solid #3eb995;
-    }
-}
-
-.startBtn {
-    background-color: #777;
-    color: #fff;
-
-    &:hover,
-    &:active {
-        background-color: #fff;
-        color: #3eb995;
-    }
-}
-
-.optInTextContainer {
-    display: flex;
-    flex-direction: column;
-    width: 110%;
-    margin-left: 10px;
-}
-
-.optInContainer {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    margin-top: 15px;
-    margin-bottom: 15px;
-}
-
 .popup {
     padding: 60px 80px;
     background: #f9fffd;
@@ -156,26 +110,6 @@
     }
 }
 
-.optIn {
-    font-size: 20px;
-    font-weight: 600;
-    color: #777;
-    text-align: center;
-    text-align: left;
-    margin-top: 0;
-    margin-bottom: 0;
-}
-
-.subOptIn {
-    font-size: 15px;
-    font-weight: 200;
-    color: #777;
-    line-height: 1em;
-    margin-bottom: 5px;
-    margin-top: 5px;
-    text-align: left;
-}
-
 .logobig {
     margin: 1.5em 0;
     width: 30%;
@@ -227,134 +161,4 @@
 p {
     margin-right: -10px;
     margin-bottom: -10px;
-}
-
-.mainContainer {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    @media (max-width: 950px) {
-        flex-direction: row;
-    }
-}
-
-.messageContainer {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    flex: 1;
-}
-
-.heading {
-    color: #3eb995;
-    font-size: 42px;
-    font-weight: 800;
-    font-style: italic;
-}
-
-.subheading {
-    font-size: 30px;
-    margin-top: 20px;
-    color: #3eb995;
-}
-
-.featureContainer {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-    flex: 1;
-    background: #fff;
-}
-
-.infoContainer {
-    display: flex;
-    position: relative;
-    align-items: center;
-    color: inherit;
-    text-decoration: none;
-    border-bottom: 1px solid #c9c9c9;
-}
-
-.infoContainer:last-child {
-    border: none;
-}
-
-.featureLink {
-    position: relative;
-    display: flex;
-    box-sizing: content-box;
-    overflow: hidden;
-    width: 100%;
-    text-decoration: none;
-    justify-content: space-between;
-    align-items: center;
-    padding: 20px 10px 20px 20px;
-    cursor: pointer;
-
-    &:hover {
-        background-color: #c2e9dd;
-    }
-}
-
-.arrow {
-    width: 40px;
-}
-
-.or {
-    margin: 80px 0;
-    position: relative;
-    left: 17%;
-    color: #3eb995;
-    font-size: 24px;
-    font-weight: 600;
-}
-
-.or::before {
-    content: '';
-    width: 15%;
-    border-bottom: solid 2px #000;
-    position: absolute;
-    top: 50%;
-    left: -17%;
-    z-index: 1;
-}
-
-.or::after {
-    content: '';
-    width: 15%;
-    border-bottom: solid 2px #000;
-    position: absolute;
-    left: 42px;
-    top: 50%;
-    z-index: 1;
-}
-
-.headingSubheading {
-    display: flex;
-    flex-direction: column;
-}
-
-.featureHeading {
-    font-weight: bold;
-    color: black;
-    text-transform: uppercase;
-}
-
-.featureSubheading {
-    color: #858585;
-    font-size: 14px;
-}
-
-.letsGoBtn {
-    display: none;
-}
-
-.startBtn:hover .getStatedBtn {
-    display: none;
-}
-
-.startBtn:hover .letsGoBtn {
-    display: block;
 }
