+++ src/overview/onboarding/components/OnboardingMsg.css
@@ -1,0 +1,103 @@
+.mainContainer {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+
+    @media (max-width: 950px) {
+        flex-direction: row;
+    }
+}
+
+.messageContainer {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    flex: 1;
+}
+
+.heading {
+    color: #3eb995;
+    font-size: 42px;
+    font-weight: 800;
+    font-style: italic;
+}
+
+.subheading {
+    font-size: 30px;
+    margin-top: 20px;
+    color: #3eb995;
+}
+
+.or {
+    margin: 80px 0;
+    position: relative;
+    left: 17%;
+    color: #3eb995;
+    font-size: 24px;
+    font-weight: 600;
+}
+
+.or::before {
+    content: '';
+    width: 15%;
+    border-bottom: solid 2px #000;
+    position: absolute;
+    top: 50%;
+    left: -17%;
+    z-index: 1;
+}
+
+.or::after {
+    content: '';
+    width: 15%;
+    border-bottom: solid 2px #000;
+    position: absolute;
+    left: 42px;
+    top: 50%;
+    z-index: 1;
+}
+
+.letsGoBtn {
+    display: none;
+}
+
+.startBtn:hover .getStatedBtn {
+    display: none;
+}
+
+.startBtn:hover .letsGoBtn {
+    display: block;
+}
+
+.choiceBtn {
+    padding: 6px 0;
+    border-radius: 2px;
+    font-weight: 600;
+    color: #fff;
+    cursor: pointer;
+    text-decoration: none;
+    background: #777;
+    border: 2px solid #777;
+    font-size: 24px;
+    text-align: center;
+    width: 234px;
+
+    &:hover,
+    &:active {
+        background-color: #fff;
+        color: #3eb995;
+        border: 2px solid #3eb995;
+    }
+}
+
+.startBtn {
+    background-color: #777;
+    color: #fff;
+
+    &:hover,
+    &:active {
+        background-color: #fff;
+        color: #3eb995;
+    }
+}
