+++ src/overview/onboarding/components/OptIn.css
@@ -1,0 +1,34 @@
+.optInTextContainer {
+    display: flex;
+    flex-direction: column;
+    width: 110%;
+    margin-left: -15px;
+}
+
+.optInContainer {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    margin-top: 15px;
+    margin-bottom: 15px;
+}
+
+.optIn {
+    font-size: 20px;
+    font-weight: 600;
+    color: #777;
+    text-align: center;
+    text-align: left;
+    margin-top: 0;
+    margin-bottom: 0;
+}
+
+.subOptIn {
+    font-size: 15px;
+    font-weight: 200;
+    color: #777;
+    line-height: 1em;
+    margin-bottom: 5px;
+    margin-top: 5px;
+    text-align: left;
+}
